refactor(analysis): move ai-validation to domain/analysis/ai-levels
백테스트 전용이던 AI SL/TP validation + ATR fallback 모듈을 production AI
analysis 파이프라인에서도 재사용할 수 있도록 domain/analysis 아래로 이동.

- src/domain/backtest/ai-validation.ts -> src/domain/analysis/ai-levels.ts
- src/__tests__/domain/backtest/ai-validation.test.ts -> src/__tests__/domain/analysis/ai-levels.test.ts
- scripts/backtests/generate-backtest.ts import path 업데이트
