+++ src/domain/analysis/ai-levels.ts
